move Rotor to rotor package and update encode()
